Split apart actions into other things.
